neterror: move IsClosed to new package
The "network" package pulls in extra dependencies like ssh and network
namespaces that the "omaha" package doesn't need. Creating a new package
for this helper function trims down dependencies for external use.
